requirements.txt
flask
pandas
scikit-learn
psycopg2-binary
SQLAlchemy
streamlit
matplotlib
requests
python-dotenv

paho-mqtt

pytest
flake8
